+++ music/.vscode/settings.json
@@ -10,7 +10,7 @@
   },
   "tslint.enable": true,
   "tslint.run": "onType",
-  "tslint.configFile": "../tslint.json",
+  "tslint.configFile": "tslint.json",
   "files.trimTrailingWhitespace": true,
   "editor.tabSize": 2,
   "editor.insertSpaces": true,
